完善简历

+++ app/src/main/java/com/zjwam/zkw/mvp/view/ICreateResumeView.java
@@ -9,5 +9,5 @@ public interface ICreateResumeView {
     void getResumeChoiceData(List<ResumePickerBean.BasicInfo> job_type,List<ResumePickerBean.BasicInfo> hiredate);
     void showMsg(String msg);
     void setProfession(List<ProfessionChoiceBean> list);
-    void saveResume(int id);
+    void saveResume(long id);
 }
